refactor(#147): finish the apidata.js split — 1,702 lines to a 64-line barrel
Completes #147. apidata.js now holds NO logic at all: only re-exports, so every
existing `from 'apidata'` import keeps working while call sites migrate to the
real module paths.

  apidata.js  1,702 -> 741 (previous passes) -> 64

## What moved in this pass

  api/globalStats.js  590 lines  the global store, rewards, donations,
                                 ArcaneOS stats, network utilisation, and
                                 fetch_global_stats itself
  api/gpuPrices.js     51 lines  GPU pricing and fleet totals
  api/walletNodes.js   +18       isWalletDOSState, a per-wallet node-health
                                 check that belongs beside the node code

Global stats was left until last deliberately: fetch_global_stats is the
orchestrator the whole app enters through, fanning out to the fluxinfo
aggregate, benchmarks, resources, geolocation, app specs and the explorer. It
touches more than any other region, and splitting it FURTHER is a design
question rather than a mechanical one -- so it moved as one unit.

Verified the 590-line block is byte-identical to what was on main. A MOVE, not
a rewrite.

isWalletDOSState went into walletNodes.js rather than a module of its own: a
fifteen-line file would be over-splitting, and it is a wallet node-health check,
which is what that module is for.

## Dead code removed

Four constants that no longer had a single reader: API_NODE_BENCHMARKS,
API_NODE_GEOLOCATION, API_FLUX_NETWORK_UTILISATION (orphaned when networkNodes
was extracted in an earlier pass) and API_DOS_LIST's duplicate.

The `appStore, StoreKeys` import, whose only two references were INSIDE a
commented-out block -- so it read as live from a grep and was not.

That commented-out Fractus helper moved to globalStats.js, beside the benchmark
data it would read. Preserved rather than deleted: Fractus is a real tier and it
documents the intended shape.

## A note on what caught what

The build caught nothing this pass, because eslint's no-unused-vars is off in
this repo (client/package.json) -- apidata.js compiled cleanly while carrying
sixteen dead imports and four dead constants. Those were found by diffing symbol
usage before and after each cut, not by tooling.

Worth knowing for the next refactor here: a clean `yarn build` is not evidence
that a file has no dead weight left in it.

## Verification

553 tests pass, production build clean, the 590-line move confirmed
byte-identical, and the image built and run.

## What is left

Nothing in apidata.js. Deleting it entirely means updating the ~15 call sites
that still import from it, which is mechanical but touches many files -- better
as its own change than bundled into this one.
